+++ dart/lib/src/client.dart
@@ -8,7 +8,6 @@ import 'client_stub.dart'
     if (dart.library.html) 'browser_client.dart'
     if (dart.library.io) 'io_client.dart';
 import 'protocol.dart';
-import 'stack_trace.dart';
 import 'utils.dart';
 import 'version.dart';
 
@@ -97,8 +96,8 @@ abstract class SentryClient {
   /// Reports an [event] to Sentry.io.
   Future<SentryId> captureEvent(
     SentryEvent event, {
-    StackFrameFilter stackFrameFilter,
     Scope scope,
+    dynamic hint,
   }) async {
     final now = options.clock();
     var authHeader = 'Sentry sentry_version=6, sentry_client=$clientId, '
@@ -127,7 +126,6 @@ abstract class SentryClient {
 
     mergeAttributes(
       event.toJson(
-        stackFrameFilter: stackFrameFilter,
         origin: origin,
       ),
       into: data,
@@ -155,13 +153,14 @@ abstract class SentryClient {
     dynamic throwable, {
     dynamic stackTrace,
     Scope scope,
+    dynamic hint,
   }) {
     final event = SentryEvent(
       exception: throwable,
       stackTrace: stackTrace,
       timestamp: options.clock(),
     );
-    return captureEvent(event, scope: scope);
+    return captureEvent(event, scope: scope, hint: hint);
   }
 
   /// Reports the [template]
@@ -171,6 +170,7 @@ abstract class SentryClient {
     String template,
     List<dynamic> params,
     Scope scope,
+    dynamic hint,
   }) {
     final event = SentryEvent(
       message: Message(
@@ -181,7 +181,7 @@ abstract class SentryClient {
       level: level,
       timestamp: options.clock(),
     );
-    return captureEvent(event, scope: scope);
+    return captureEvent(event, scope: scope, hint: hint);
   }
 
   void close() {