+++ dart/lib/src/noop_client.dart
@@ -34,11 +34,20 @@ class NoOpSentryClient implements SentryClient {
   Map<String, String> buildHeaders(String authHeader) => {};
 
   @override
-  Future<SentryId> captureEvent(SentryEvent event, {stackFrameFilter, scope}) =>
+  Future<SentryId> captureEvent(
+    SentryEvent event, {
+    Scope scope,
+    dynamic hint,
+  }) =>
       Future.value(SentryId.empty());
 
   @override
-  Future<SentryId> captureException(throwable, {stackTrace, scope}) =>
+  Future<SentryId> captureException(
+    throwable, {
+    dynamic stackTrace,
+    Scope scope,
+    dynamic hint,
+  }) =>
       Future.value(SentryId.empty());
 
   @override
@@ -48,6 +57,7 @@ class NoOpSentryClient implements SentryClient {
     String template,
     List<dynamic> params,
     Scope scope,
+    dynamic hint,
   }) =>
       Future.value(SentryId.empty());