+++ dart/test/sentry_test.dart
@@ -27,7 +27,6 @@ void main() {
         client.captureEvent(
           fakeEvent,
           scope: anyNamed('scope'),
-          stackFrameFilter: null,
         ),
       ).called(1);
     });